Refactor/db currency type

+++ proap-api/src/main/java/br/ufba/proap/assistancerequest/domain/dto/AssistanceRequestCeapgDTO.java
@@ -1,10 +1,12 @@
 package br.ufba.proap.assistancerequest.domain.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class AssistanceRequestCeapgDTO {
 
-    private Float custoFinalCeapg;
+    private BigDecimal custoFinalCeapg;
     private String observacoesCeapg;
 }
